helm

+++ Jenkinsfile
@@ -112,8 +112,10 @@ pipeline {
                 script {
                     sh "docker rmi ${dockerRepoName}:latest"
                     sh "docker rmi ${dockerRepoName}:${buildName}"
-                    sh "docker rmi ${dockerRegistryNoProto}/${dockerRepoName}:${buildName}"
-                    sh "docker rmi ${dockerRegistryNoProto}/${dockerRepoName}:latest"
+                    if (params.deployEnv == "${stagingEnv}") {
+                        sh "docker rmi ${dockerRegistryNoProto}/${dockerRepoName}:${buildName}"
+                        sh "docker rmi ${dockerRegistryNoProto}/${dockerRepoName}:latest"
+                    }
                     sh "yes | docker image prune"
                 }
             }
